// frontend/src/interfaces/tables.interfaces.ts
export interface IFcrbTableRow {
	accruedInterest: number
	benefitInMonth: number
	deposit202: number
	deposit204: number
	deposit206: number
	filialName: string
	issuedLoans: number
	mfo: string
	npl: number
	nplPercent: number
	par30: number
	par60: number
	par90: number
	resourceDebt: number
	roa: number
	roe: number
	totalLoan: number
}